Clone Page - Microsoft
Microsoft is a technology company founded by Bill Gates and Paul Allen in 1975 in the United States. Headquartered in Redmond (Washington), the company operates in the support and development of software, devices and services related to personal computing, productivity and cloud solutions. Furthermore, the brand also has significant participation in the artificial intelligence, virtual reality and games segments.

Some achievements mark the beginning of its trajectory: the first product developed by the company was a version of the BASIC interpreter, for the MITS Altair 8800 computer. In 1976, Microsoft FORTRAN was released for CP/M-based computers. It is worth mentioning that the company entered the operating systems business in 1980, with its own version of Unix, called Xenix. However, in 1985 a new page was written in history when Microsoft launched Windows.




THIS IS JUST A CLONE PAGE, WITH THE INTEND TO GET MORE KNOWLEDGE IN THE AREA OF PROGRAMMING